revive ptrace #5983 HW CI
