+++ java_console/autotest/src/main/java/com/rusefi/HwCiProteus.java
@@ -3,6 +3,7 @@ package com.rusefi;
 import com.rusefi.common.MiscTest;
 import com.rusefi.f4discovery.CompositeLoggerTest;
 import com.rusefi.f4discovery.HighRevTest;
+import com.rusefi.f4discovery.PTraceTest;
 import com.rusefi.proteus.ProteusAnalogTest;
 
 /**
@@ -11,6 +12,7 @@ import com.rusefi.proteus.ProteusAnalogTest;
 public class HwCiProteus {
     public static void main(String[] args) {
         CmdJUnitRunner.runHardwareTestAndExit(new Class[]{
+            PTraceTest.class,
                 CompositeLoggerTest.class,
                 HighRevTest.class,
                 MiscTest.class,
